update to sync-exec for node 0.12+
According to this issue: https://github.com/mgutz/execSync/issues/38 - node 0.12+ is not working with execSync. Since an update cannot be pushed because of the camel-cased name, a new version of was created here: https://www.npmjs.com/package/sync-exec

I've updated the package.json to reference this new version as I was unable to generate using Yeoman because of this issue.